requirements.txt
Flask==2.3.3
reportlab==4.0.4
psycopg2-binary==2.9.9
Flask-Login==0.6.3
Authlib==1.2.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
pytest-mock==3.12.0
